[Feature #163550153] Refactor code
